feat: Add student mode and references to prescriptions
Introduces a student mode to the application, disabling legal value for generated documents and providing an academic context. This mode is visually indicated by a banner in the preview.

Additionally, a new `reference` field has been added to the `PrescriptionState` type to allow for storing guideline or source references for prescriptions.